Add Clear_Keyboard & Clear_Pads

+++ WindowsFormsAppMusicPad/TrackName.cs
@@ -7,9 +7,9 @@ namespace WindowsFormsAppMusicPad
         public string path { get; set; }
         public string name { get; set; }
         public string ButtonName { get; set; }
-        public byte note { get; set; }
+        public int note { get; set; }
 
-        public TrackName(string pth, string butName, byte note)
+        public TrackName(string pth, string butName, int note)
         {
             path = pth; 
             name = Path.GetFileNameWithoutExtension(pth);
